chore: remove legacy Turbo/Stimulus artifacts
Drop unused Hotwire gems, Stimulus controllers, importmap pins, and
backup ERB views left after the React migration. Replace the Turbo
category filter system test with React DOM selectors.

Closes #190
