+++ Gemfile
@@ -1,4 +1,4 @@
-source "https://rubygems.org"
+﻿source "https://rubygems.org"
 
 # Bundle edge Rails instead: gem "rails", github: "rails/rails", branch: "main"
 gem "rails", "~> 8.1.3"
@@ -10,7 +10,6 @@ gem "pg", "~> 1.6"
 gem "puma", ">= 5.0"
 # Use Vite for frontend assets [https://github.com/ElMassimo/vite_ruby]
 gem "vite_rails"
-gem "turbo-rails"
 # Build JSON APIs with ease [https://github.com/rails/jbuilder]
 gem "jbuilder"
