+++ app/views/layouts/application.html.erb
@@ -19,7 +19,7 @@
     <link rel="apple-touch-icon" href="/icon.png">
 
     <%# Includes all stylesheet files in app/assets/stylesheets %>
-    <%= stylesheet_link_tag :app, "data-turbo-track": "reload" %>
+    <%= stylesheet_link_tag :app %>
     <%= vite_client_tag %>
     <%# Link CSS as its own entry so styles load with built assets (no Vite HMR). %>
     <%= vite_stylesheet_tag 'application.css' %>
