--- bin/importmap
@@ -1,4 +1,0 @@
-#!/usr/bin/env ruby
-
-require_relative "../config/application"
-require "importmap/commands"
